covidNews/searchNews/requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
dataclasses==0.7
Django==3.0.7
django-crontab==0.7.1
elasticsearch==7.7.1
goose3==3.1.6
idna==2.9
jieba==0.42.1
joblib==0.15.1
lxml==4.5.1
nltk==3.5
numpy==1.19.0
oauthlib==3.1.0
pandas==1.0.5
Pillow==7.1.2
pkg-resources==0.0.0
preprocessor==1.1.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
regex==2020.6.8
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
textblob==0.15.3
tqdm==4.46.1
tweepy==3.9.0
urllib3==1.25.9
